requirements.txt
argparse==1.2.1
dj-database-url==0.2.1
django-compressor==1.3
boto==2.9.6
git+https://github.com/django-debug-toolbar/django-debug-toolbar.git#egg=django-debug-toolbar
django-filter==0.6
django-countries==2.0b4
django-registration==1.0
django-storages==1.1.8
djangorestframework==2.3.12
Django==1.6.1
djrill==0.7.0
mimeparse==0.1.3
psycopg2==2.4.6
South==0.8.4
Pillow==2.3.0
sorl-thumbnail==11.12
docutils==0.11
markdown==2.3.1
gunicorn==0.17.4
